01/18

+++ src/main/java/jdbcboard/service/BoardServiceImpl.java
@@ -17,22 +17,35 @@ public class BoardServiceImpl implements BoardService{
     public List<Board> getBoards(int page) {
         List<Board> boards = new ArrayList<>();
         BoardDao boardDao = new BoardDaoImpl();
-        Connection conn = null;
 
         int start = SIZE * page - SIZE;
         int limit = SIZE;
 
-        try {
-            conn = DBUtil.getInstance().getConnection();
+        try (Connection conn = DBUtil.getInstance().getConnection();){
             ConnectionContextHolder.setConnection(conn);
             boards = boardDao.getBoards(start, limit);
         }catch(Exception ex){
             ex.printStackTrace();
-        }finally {
-            DBUtil.close(conn);
         }
 
         return boards;
+    }
+
+    public long getBoardCount() {
+        long count = 0L;
+        BoardDao boardDao = new BoardDaoImpl();
+        Connection conn = null;
+        try{
+            conn = DBUtil.getInstance().getConnection();
+            ConnectionContextHolder.setConnection(conn);
+            count = boardDao.getCount();
+        }catch(Exception ex){
+            DBUtil.rollback(conn);
+            ex.printStackTrace();
+        }finally {
+            DBUtil.close(conn);
+        }
+        return count;
     }
 
     @Override
@@ -57,13 +70,24 @@ public class BoardServiceImpl implements BoardService{
     }
 
     @Override
-    public void deleteBoard(Long id) {
+    public void deleteBoard(Long id, Long sigendId) {
         BoardDao boardDao= new BoardDaoImpl();
         Connection conn = null;
         try{
             conn = DBUtil.getInstance().getConnection();
             ConnectionContextHolder.setConnection(conn);
-            boardDao.deleteBoard(id);
+
+            Board board = boardDao.getBoard(id);
+            if(board.getUserId()==sigendId){
+                if(!boardDao.existReply(board.getThread())){
+                    boardDao.deleteBoard(id);
+                    boardDao.updateCountBoardMinus();
+                    if(board.getThread() == boardDao.getMaxThread()*100){
+                        boardDao.updateMaxThreadMinus();
+                    }
+                }
+            }
+
             conn.commit(); // 트랜젝션 commit
         }catch (Exception ex){
             DBUtil.rollback(conn);
@@ -80,7 +104,11 @@ public class BoardServiceImpl implements BoardService{
         try{
             conn = DBUtil.getInstance().getConnection();
             ConnectionContextHolder.setConnection(conn);
+            // board 추가.
             boardDao.addBoard(board);
+            // MAX THREAD + 1 , COUNT BOARD + 1
+            boardDao.updateMaxThread();
+            boardDao.updateCountBoard();
             conn.commit(); // 트랜젝션 commit
         }catch (Exception ex){
             DBUtil.rollback(conn);
@@ -98,6 +126,34 @@ public class BoardServiceImpl implements BoardService{
             conn = DBUtil.getInstance().getConnection();
             ConnectionContextHolder.setConnection(conn);
             boardDao.updateBoard(board);
+            conn.commit(); // 트랜젝션 commit
+        }catch (Exception ex){
+            DBUtil.rollback(conn);
+            ex.printStackTrace();
+        }finally {
+            DBUtil.close(conn);
+        }
+    }
+
+    @Override
+    public void addReply(long parentId, Board board) {
+        BoardDao boardDao= new BoardDaoImpl();
+        Connection conn = null;
+        try{
+            conn = DBUtil.getInstance().getConnection();
+            ConnectionContextHolder.setConnection(conn);
+            // 1. 부모 thread, depth 를 가져온다.
+            Board tmpBoard = boardDao.getThreadDepth(parentId);
+            Long thread = tmpBoard.getThread();
+            int depth = tmpBoard.getDepth();
+            // 2. thread 업데이트. ( thread unique??...)
+            boardDao.updateThreadMinus(thread);
+            // 3. board 설정
+            board.setThread(thread - 1);
+            board.setDepth(depth + 1);
+            boardDao.addReply(board);
+            // 4. manage 업데이트.
+            boardDao.updateCountBoard();
             conn.commit(); // 트랜젝션 commit
         }catch (Exception ex){
             DBUtil.rollback(conn);
